shakecraft/app/requirements.txt
gunicorn==20.1.0
Flask==2.0.3
Jinja2==3.0.3
itsdangerous==2.1.1
Werkzeug==2.0.3
click==8.0.4
WTForms==3.0.1
MarkupSafe==2.1.1
Flask-WTF==1.0.0
tensorflow-cpu==2.9.1
